Meld two implementations into a single one
